chore: gardenCommandService의 메서드를 역할에 맞게 분배
