+++ src/main/java/doore/study/api/CurriculumItemController.java
@@ -1,6 +1,5 @@
 package doore.study.api;
 
-import doore.garden.application.GardenCommandService;
 import doore.study.application.CurriculumItemCommandService;
 import doore.study.application.dto.request.CurriculumItemManageRequest;
 import jakarta.validation.Valid;
@@ -20,7 +19,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class CurriculumItemController {
 
     private final CurriculumItemCommandService curriculumItemCommandService;
-    private final GardenCommandService gardenCommandService;
 
     @PostMapping("/studies/{studyId}/curriculums")
     public ResponseEntity<Void> manageCurriculum(@PathVariable Long studyId,
