+++ src/main/java/doore/study/application/CurriculumItemCommandService.java
@@ -6,8 +6,9 @@ import static doore.study.exception.CurriculumItemExceptionType.NOT_FOUND_CURRIC
 import static doore.study.exception.StudyExceptionType.NOT_FOUND_PARTICIPANT;
 import static doore.study.exception.StudyExceptionType.NOT_FOUND_STUDY;
 
-import doore.garden.application.GardenCommandService;
+import doore.garden.domain.Garden;
 import doore.garden.domain.GardenType;
+import doore.garden.domain.repository.GardenRepository;
 import doore.member.domain.Participant;
 import doore.member.domain.repository.ParticipantRepository;
 import doore.study.application.dto.request.CurriculumItemManageDetailRequest;
@@ -37,7 +38,7 @@ public class CurriculumItemCommandService {
     private final ParticipantCurriculumItemRepository participantCurriculumItemRepository;
     private final StudyRepository studyRepository;
     private final ParticipantRepository participantRepository;
-    private final GardenCommandService gardenCommandService;
+    private final GardenRepository gardenRepository;
 
     public void manageCurriculum(CurriculumItemManageRequest request, Long studyId) {
         List<CurriculumItemManageDetailRequest> curriculumItems = request.curriculumItems();
@@ -61,10 +62,22 @@ public class CurriculumItemCommandService {
 
         participantCurriculumItem.checkCompletion();
         if (participantCurriculumItem.getIsChecked()) {
-            gardenCommandService.deleteGarden(participantCurriculumItem);
+            deleteGarden(participantCurriculumItem);
             return;
         }
-        gardenCommandService.createGarden(participantCurriculumItem);
+        createGarden(participantCurriculumItem);
+    }
+
+    public void createGarden(ParticipantCurriculumItem participantCurriculumItem) {
+        Garden garden = GardenType.getSupplierOf(participantCurriculumItem.getClass().getSimpleName())
+                .of(participantCurriculumItem);
+        gardenRepository.save(garden);
+    }
+
+    public void deleteGarden(ParticipantCurriculumItem participantCurriculumItem) {
+        Long contributionId = participantCurriculumItem.getId();
+        GardenType gardenType = GardenType.getGardenTypeOf(participantCurriculumItem.getClass().getSimpleName());
+        gardenRepository.deleteByContributionIdAndType(contributionId, gardenType);
     }
 
     private void checkItemOrderDuplicate(List<CurriculumItemManageDetailRequest> curriculumItems) {
