+++ src/test/java/doore/garden/application/GardenCommandServiceTest.java
@@ -29,9 +29,6 @@ public class GardenCommandServiceTest extends IntegrationTest {
     DocumentCommandService documentCommandService;
 
     @Autowired
-    GardenCommandService gardenCommandService;
-
-    @Autowired
     ParticipantCurriculumItemRepository participantCurriculumItemRepository;
     @Autowired
     GardenRepository gardenRepository;
@@ -39,59 +36,7 @@ public class GardenCommandServiceTest extends IntegrationTest {
     CurriculumItemRepository curriculumItemRepository;
 
 
-    @Test
-    @DisplayName("[성공] 커리큘럼 체크시 정상적으로 텃밭을 생성할 수 있다.")
-    public void createGarden_커리큘럼_체크시_정상적으로_텃밭에_반영된다_성공() throws Exception {
-        //given
-        CurriculumItem curriculumItem = curriculumItemRepository.save(curriculumItem());
-        ParticipantCurriculumItem participantCurriculumItem = ParticipantCurriculumItem.builder()
-                .participantId(1L)
-                .curriculumItem(curriculumItem)
-                .build();
-        participantCurriculumItemRepository.save(participantCurriculumItem);
 
-        //when
-        gardenCommandService.createGarden(participantCurriculumItem);
 
-        //then
-        Garden garden = gardenRepository.findAll().get(0);
-        assertEquals(garden.getContributionId(), participantCurriculumItem.getId());
-        assertEquals(garden.getType(), STUDY_CURRICULUM_COMPLETION);
-    }
 
-    @Test
-    @DisplayName("[성공] 학습자료 업로드시 정상적으로 텃밭을 생성할 수 있다.")
-    public void createGarden_학습자료_업로드시_정상적으로_텃밭에_반영된다_성공() throws Exception {
-        //given
-        Document document = new DocumentFixture().buildDocument();
-
-        //when
-        gardenCommandService.createGarden(document);
-
-        //then
-        Garden garden = gardenRepository.findAll().get(0);
-        assertEquals(garden.getContributionId(), document.getId());
-        assertEquals(garden.getType(), DOCUMENT_UPLOAD);
-    }
-
-    @Test
-    @DisplayName("[성공] 커리큘럼 체크 해재시 정상적으로 텃밭에 반영된다.")
-    public void deleteGarden_커리큘럼_체크_해재시_정상적으로_텃밭에_반영된다_성공() throws Exception {
-        //given
-        CurriculumItem curriculumItem = curriculumItemRepository.save(curriculumItem());
-        ParticipantCurriculumItem participantCurriculumItem = ParticipantCurriculumItem.builder()
-                .participantId(1L)
-                .curriculumItem(curriculumItem)
-                .build();
-        participantCurriculumItemRepository.save(participantCurriculumItem);
-
-        gardenCommandService.createGarden(participantCurriculumItem);
-        assertEquals(1, gardenRepository.findAll().size());
-
-        //when
-        gardenCommandService.deleteGarden(participantCurriculumItem);
-
-        //then
-        assertEquals(0, gardenRepository.findAll().size());
-    }
 }
